+++ common/sdk/CHANGELOG.md
@@ -3,13 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [1.0.0-beta.10](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/sdk@1.0.0-beta.9...@kyvejs/sdk@1.0.0-beta.10) (2023-03-13)
+# [1.0.0-beta.11](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/sdk@1.0.0-beta.10...@kyvejs/sdk@1.0.0-beta.11) (2023-03-22)
 
 **Note:** Version bump only for package @kyvejs/sdk
 
+# [1.0.0-beta.10](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/sdk@1.0.0-beta.9...@kyvejs/sdk@1.0.0-beta.10) (2023-03-13)
 
-
-
+**Note:** Version bump only for package @kyvejs/sdk
 
 # [1.0.0-beta.9](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/sdk@1.0.0-beta.8...@kyvejs/sdk@1.0.0-beta.9) (2023-03-13)