+++ integrations/uniswap/CHANGELOG.md
@@ -3,13 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [1.0.0-beta.14](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/uniswap@1.0.0-beta.13...@kyvejs/uniswap@1.0.0-beta.14) (2023-03-13)
+# [1.0.0-beta.16](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/uniswap@1.0.0-beta.15...@kyvejs/uniswap@1.0.0-beta.16) (2023-03-31)
 
 **Note:** Version bump only for package @kyvejs/uniswap
 
+# [1.0.0-beta.15](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/uniswap@1.0.0-beta.14...@kyvejs/uniswap@1.0.0-beta.15) (2023-03-22)
 
+**Note:** Version bump only for package @kyvejs/uniswap
 
+# [1.0.0-beta.14](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/uniswap@1.0.0-beta.13...@kyvejs/uniswap@1.0.0-beta.14) (2023-03-13)
 
+**Note:** Version bump only for package @kyvejs/uniswap
 
 # [1.0.0-beta.13](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/uniswap@1.0.0-beta.12...@kyvejs/uniswap@1.0.0-beta.13) (2023-03-13)